Refactor manifest and kube converter
* split jobs and instance groups into separate files
* kube converter is now KubeConfig.Convert
* try to define methods on the structs they'll modify
